improve gdb
